requirements.txt
PyYAML
pytest
ruff
black